+++ packages/next-common/components/multisigs/desktop/index.js
@@ -6,15 +6,9 @@ import {
   myMultisigsSelector,
 } from "next-common/store/reducers/multisigSlice";
 import { AddressUser } from "next-common/components/user";
-import {
-  Approving,
-  Call,
-  Signatories,
-  Status,
-  When,
-  SignStatus,
-} from "../fields";
+import { Approving, Call, Signatories, Status, When } from "../fields";
 import columns from "./colmns";
+import MultisigSignField from "../signField";
 
 export default function DesktopList() {
   const myMultisigs = useSelector(myMultisigsSelector);
@@ -46,7 +40,7 @@ export default function DesktopList() {
       args={multisig.state.args}
       updateAt={multisig.updateAt}
     />,
-    <SignStatus key="signStatus" multisig={multisig} />,
+    <MultisigSignField key="signStatus" multisig={multisig} />,
   ]);
 
   return (
